+++ SGGL-Diablo-Knowledge-Library/src/helper/encoding.h
