+++ SGGL-Diablo-Knowledge-Library/src/helper/error_handling.c
